resumeParser/requirements.txt
annotated-types
anyio
blinker
certifi
click
colorama
distro
exceptiongroup
Flask
Flask-Cors
h11
httpcore
httpx
idna
itsdangerous
Jinja2
jiter
Markdown
MarkupSafe
openai
pydantic
pydantic_core
sniffio
tqdm
typing_extensions
Werkzeug
PyPDF2
python-dotenv